Extract Hello::App::Router into Wren::Router
The first wren module is born!
